create db.py file for database configurations
